adding Timetable entity in order to regroup common methods/attributes between Stop and Line timetables.
