+++ Entity/StopTimetable.php
@@ -5,7 +5,7 @@ namespace CanalTP\MttBundle\Entity;
 /**
  * StopTimetable
  */
-class StopTimetable extends AbstractEntity
+class StopTimetable extends Timetable
 {
     /**
      * @var integer
@@ -16,11 +16,6 @@ class StopTimetable extends AbstractEntity
      * @var string
      */
     private $externalRouteId;
-
-    /**
-     * @var Object
-     */
-    private $blocks;
 
     /**
      * @var Object
@@ -78,29 +73,6 @@ class StopTimetable extends AbstractEntity
     public function getExternalRouteId()
     {
         return $this->externalRouteId;
-    }
-
-    /**
-     * Set blocks
-     *
-     * @param  array $blocks
-     * @return Line
-     */
-    public function setBlocks($blocks)
-    {
-        $this->blocks = $blocks;
-
-        return $this;
-    }
-
-    /**
-     * Get blocks
-     *
-     * @return array
-     */
-    public function getBlocks()
-    {
-        return $this->blocks;
     }
 
     /**
@@ -216,10 +188,5 @@ class StopTimetable extends AbstractEntity
     public function getLineConfig()
     {
         return $this->line_config;
-    }
-
-    public function isLocked()
-    {
-        return $this->getLineConfig()->isLocked();
     }
 }
